MAILET-115 Remove abstract service with original mail implementation
