+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/redirect/AbstractRedirect.java
@@ -23,11 +23,8 @@ import java.util.List;
 
 import javax.inject.Inject;
 import javax.mail.MessagingException;
-import javax.mail.Session;
 import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeMessage;
 
-import org.apache.james.core.MailImpl;
 import org.apache.james.dnsservice.api.DNSService;
 import org.apache.james.transport.mailets.Redirect;
 import org.apache.james.transport.mailets.utils.MimeMessageModifier;
@@ -313,125 +310,7 @@ public abstract class AbstractRedirect extends GenericMailet {
      * @throws MessagingException if a problem arises formulating the redirected mail
      */
     @Override
-    public void service(Mail originalMail) throws MessagingException {
-
-        boolean keepMessageId = false;
-
-        // duplicates the Mail object, to be able to modify the new mail keeping
-        // the original untouched
-        MailImpl newMail = new MailImpl(originalMail);
-        try {
-            MailModifier mailModifier = MailModifier.builder()
-                    .mailet(this)
-                    .mail(newMail)
-                    .dns(dns)
-                    .build();
-            mailModifier.setRemoteAddr();
-            mailModifier.setRemoteHost();
-
-            if (getInitParameters().isDebug()) {
-                log("New mail - sender: " + newMail.getSender() + ", recipients: " + arrayToString(newMail.getRecipients().toArray()) + ", name: " + newMail.getName() + ", remoteHost: " + newMail.getRemoteHost() + ", remoteAddr: " + newMail.getRemoteAddr() + ", state: " + newMail.getState()
-                        + ", lastUpdated: " + newMail.getLastUpdated() + ", errorMessage: " + newMail.getErrorMessage());
-            }
-
-            // Create the message
-            if (!getInitParameters().getInLineType().equals(TypeCode.UNALTERED)) {
-                if (getInitParameters().isDebug()) {
-                    log("Alter message");
-                }
-                newMail.setMessage(new MimeMessage(Session.getDefaultInstance(System.getProperties(), null)));
-
-                // handle the new message if altered
-                AlteredMailUtils.builder()
-                    .mailet(this)
-                    .originalMail(originalMail)
-                    .build()
-                    .buildAlteredMessage(newMail);
-
-            } else {
-                // if we need the original, create a copy of this message to
-                // redirect
-                if (getInitParameters().getPassThrough()) {
-                    newMail.setMessage(new MimeMessage(originalMail.getMessage()) {
-                        protected void updateHeaders() throws MessagingException {
-                            if (getMessageID() == null)
-                                super.updateHeaders();
-                            else {
-                                modified = false;
-                            }
-                        }
-                    });
-                }
-                if (getInitParameters().isDebug()) {
-                    log("Message resent unaltered.");
-                }
-                keepMessageId = true;
-            }
-
-            // Set additional headers
-
-            mailModifier.setRecipients(getRecipients(originalMail));
-            mailModifier.setTo(getTo(originalMail));
-            mailModifier.setSubjectPrefix(originalMail);
-            mailModifier.setReplyTo(getReplyTo(originalMail), originalMail);
-            mailModifier.setReversePath(getReversePath(originalMail), originalMail);
-            mailModifier.setIsReply(getInitParameters().isReply(), originalMail);
-            mailModifier.setSender(getSender(originalMail), originalMail);
-            mailModifier.initializeDateIfNotPresent();
-            if (keepMessageId) {
-                mailModifier.setMessageId(originalMail);
-            }
-            newMail =  mailModifier.getMail();
-
-            newMail.getMessage().saveChanges();
-            newMail.removeAllAttributes();
-
-            if (senderDomainIsValid(newMail)) {
-                // Send it off...
-                getMailetContext().sendMail(newMail);
-            } else {
-                String logBuffer = getMailetName() + " mailet cannot forward " + originalMail.getName() + ". Invalid sender domain for " + newMail.getSender() + ". Consider using the Resend mailet " + "using a different sender.";
-                throw new MessagingException(logBuffer);
-            }
-
-        } finally {
-            newMail.dispose();
-        }
-
-        if (!getInitParameters().getPassThrough()) {
-            originalMail.setState(Mail.GHOST);
-        }
-    }
+    public abstract void service(Mail originalMail) throws MessagingException;
 
     protected abstract MimeMessageModifier getMimeMessageModifier(Mail newMail, Mail originalMail) throws MessagingException;
-
-    /**
-     * <p>
-     * Checks if a sender domain of <i>mail</i> is valid.
-     * </p>
-     * <p>
-     * If we do not do this check, and someone uses a redirection mailet in a
-     * processor initiated by SenderInFakeDomain, then a fake sender domain will
-     * cause an infinite loop (the forwarded e-mail still appears to come from a
-     * fake domain).<br>
-     * Although this can be viewed as a configuration error, the consequences of
-     * such a mis-configuration are severe enough to warrant protecting against
-     * the infinite loop.
-     * </p>
-     * <p>
-     * This check can be skipped if {@link #getFakeDomainCheck(Mail)} returns
-     * true.
-     * </p>
-     *
-     * @param mail the mail object to check
-     * @return true if the if the sender is null or
-     *         {@link org.apache.mailet.MailetContext#getMailServers} returns
-     *         true for the sender host part
-     */
-    @SuppressWarnings("deprecation")
-    protected final boolean senderDomainIsValid(Mail mail) throws MessagingException {
-        return !getInitParameters().getFakeDomainCheck()
-                || mail.getSender() == null
-                || !getMailetContext().getMailServers(mail.getSender().getDomain()).isEmpty();
-    }
 }
